+++ NEWS.md
@@ -1,3 +1,10 @@
+# Changes in version 1.1.2
+
+* adding volume instead of area for MERFISH.
+* refactoring log2CountArea to log2DensitySignal
+* added size and alpha arguments to plotCellsFovs
+* added scaleBar argument to plotCellsFovs, plotCentroids, plotPolygons and plotZoomFovsMap
+
 # Changes in version 1.0.1
 
 * fixing minor bugs in AspectRatio internal computation for technology missing it.